requirements.txt
python-docx==0.8.11
pydantic>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pydantic-core>=2.0.0